+++ huey/registry.py
@@ -7,7 +7,7 @@ from huey.utils import ChordConfig
 Message = namedtuple('Message', ('id', 'name', 'eta', 'retries', 'retry_delay',
                                  'priority', 'args', 'kwargs', 'on_complete',
                                  'on_error', 'expires', 'expires_resolved',
-                                 'timeout', 'chord_config'))
+                                 'timeout', 'chord_config', 'retry_backoff'))
 
 # Automatically set missing parameters to None. This is kind-of a hack, but it
 # allows us to add new parameters while continuing to be able to handle
@@ -91,7 +91,8 @@ class Registry(object):
             task.expires,
             task.expires_resolved,
             task.timeout,
-            chord_config)
+            chord_config,
+            task.retry_backoff)
 
     def create_task(self, message):
         TaskClass = self.string_to_task(message.name)
@@ -123,7 +124,8 @@ class Registry(object):
             on_error,
             message.expires_resolved,
             message.timeout,
-            chord_config)
+            chord_config,
+            message.retry_backoff)
 
     @property
     def periodic_tasks(self):